[KOGITO-7524] Tiago suggestions for refactor
