+++ api/kogito-api/src/main/java/org/kie/kogito/internal/process/event/KogitoObjectListenerAware.java
@@ -19,6 +19,15 @@ import java.util.Collection;
 
 public interface KogitoObjectListenerAware {
 
+    default void fireEvent(String propertyName, Object oldValue, Object newValue, Runnable updater) {
+        if (newValue instanceof KogitoObjectListenerAware) {
+            ((KogitoObjectListenerAware) newValue).addKogitoObjectListener(new KogitoObjectListenerAwareListener(this, propertyName));
+        }
+        listeners().forEach(l -> l.beforeValueChanged(this, propertyName, oldValue, newValue));
+        updater.run();
+        listeners().forEach(l -> l.afterValueChanged(this, propertyName, oldValue, newValue));
+    }
+
     void addKogitoObjectListener(KogitoObjectListener listener);
 
     Collection<KogitoObjectListener> listeners();
